Fixed bugs
I just can't remember what they are.

+++ src/io/thought/Searcher.java
@@ -1,5 +1,5 @@
 package io.thought;
 
 public interface Searcher {
-	public int getDistance(int[] currentNumbs, int index, int value, int[] parameters) throws InvalidNumberException;
+	public long getDistance(int[] currentNumbs, int index, int value, int[] parameters) throws InvalidNumberException;
 }
